Fixed tests to better reflect new changes
