+++ conference/forms/talks.py
@@ -6,9 +6,7 @@ from taggit_labels.widgets import LabelWidget
 
 from conference.forms import TalkBaseForm
 from conference.models import (Conference, ConferenceTag, Talk, TALK_TYPE,
-    CFP_TALK_TYPE, AttendeeProfile)
-
-from phonenumber_field.formfields import PhoneNumberField
+    CFP_TALK_TYPE)
 
 
 class TalkUpdateForm(forms.ModelForm):
@@ -63,70 +61,6 @@ class TalkUpdateForm(forms.ModelForm):
 
         return talk
 
-class AddSpeakerToTalkForm(forms.ModelForm):
-    users_given_name = forms.CharField(label="Given name of the speaker")
-    users_family_name = forms.CharField(label="Family name of the speaker")
-    is_minor = forms.BooleanField(
-        label="Are you a minor?",
-        help_text=(
-            "Please select this checkbox if you're going to be under 18"
-            "years old on July 26th 2021"
-        ),
-        # required=False, because django forms... it means that unless someone
-        # is a minor we don't provide a value.
-        required=False,
-    )
-    job_title = forms.CharField(
-        label="Job title",
-        help_text="eg: student, developer, CTO, js ninja, BDFL",
-        max_length=50,
-        required=False,
-    )
-    phone = PhoneNumberField(
-        help_text=(
-            "We require a mobile phone number for all speakers "
-            "for last minute contacts and in case we need "
-            "timely clarification (if no reponse to previous emails). "
-            "Use the international format (e.g.: +44 123456789). "
-            "This field will <strong>never</strong> be published."
-        ),
-        max_length=30,
-    )
-    company = forms.CharField(
-        label="Your company", max_length=50, required=False
-    )
-    company_homepage = forms.URLField(
-        label="Company homepage", required=False
-    )
-    bio = forms.CharField(
-        label="Compact biography",
-        help_text=(
-            "Please enter a short biography (one or two paragraphs) <br />"
-            "Do not paste your CV!"
-        ),
-        widget=forms.Textarea(),
-    )
-    i_accept_speaker_release  = TalkBaseForm.base_fields[
-        'i_accept_speaker_release'
-    ]
-
-
-    class Meta:
-        model = AttendeeProfile
-        fields = [
-            'users_given_name',
-            'users_family_name',
-            'job_title',
-            'is_minor',
-            'phone',
-            'bio',
-            'company',
-            'company_homepage',
-        ]
-
-
-class UpdateAttendeeProfile(AddSpeakerToTalkForm):
-    pass
 
 class TalkSlidesForm(forms.ModelForm):
     class Meta:
